fix(keyboard) repair ipc communication
Also set AZERTY layout (not configurable yet)
