+++ src/config/preload.ts
@@ -1,26 +1,3 @@
-import { contextBridge, IpcRendererEvent, ipcRenderer } from 'electron'
+import { exposeIpc } from '../common/exposeIpc'
 
-const enabledEvents = ['close', 'getAppConfig', 'receiveConfig', 'setAppConfig']
-
-function filterEnabledEvent(name: string) {
-    if (!enabledEvents.includes(name)) {
-        throw Error(`Event ${name} is not supported`)
-    }
-}
-
-const ipcBridge: IpcBridge = {
-    on: (name: string, callback: (evt: IpcRendererEvent, ...values: any[]) => void) => {
-        filterEnabledEvent(name)
-        ipcRenderer.on(name, callback)
-    },
-    off: (name: string, callback: (evt: IpcRendererEvent, ...values: any[]) => void) => {
-        filterEnabledEvent(name)
-        ipcRenderer.off(name, callback)
-    },
-    send: (name: string, ...values: any[]) => {
-        filterEnabledEvent(name)
-        ipcRenderer.send(name, ...values)
-    }
-}
-
-contextBridge.exposeInMainWorld('ipc', ipcBridge)
+exposeIpc(['close', 'getAppConfig', 'receiveConfig', 'setAppConfig'])
